requirements.txt
Flask>=2.3
pandas>=2.0
numpy>=1.24
joblib>=1.3
python-dotenv>=1.0
google-generativeai>=0.8
requests>=2.31
beautifulsoup4>=4.12
lxml>=5.2
rapidfuzz>=3.9
langchain>=0.3
langchain-community>=0.3
langchain-core>=0.3
SQLAlchemy>=2.0
